--- scripts/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "../tsconfig.json",
-	"compilerOptions": {
-		"allowJs": true,
-		"moduleResolution": "Node"
-	},
-	"include": ["./**/*.mjs", "./**/*.ts"]
-}
